+++ entity/product-category.go
@@ -1,7 +1,7 @@
 package entity
 
 type ProductCategory struct {
-	ID      int8   `json:"id"`
+	ID      int    `json:"id"`
 	Name    string `json:"name"`
 	StoreId string `json:"store_id"`
 }
