+++ entity/product.go
@@ -1,9 +1,9 @@
 package entity
 
 type Product struct {
-	ID                int8    `json:"id"`
+	ID                int     `json:"id"`
 	Name              string  `json:"name"`
-	StoreId           int8    `json:"store_id"`
+	StoreId           int     `json:"store_id"`
 	StoreName         string  `json:"store_name"`
 	ProductCategoryId string  `json:"product_category_id"`
 	Image             string  `json:"image"`
