+++ entity/user.go
@@ -1,16 +1,16 @@
 package entity
 
 type User struct {
-	ID        int8      `json:"id"`
+	ID        int       `json:"id"`
 	Name      string    `json:"name" binding:"required"`     /*,min:6,max=50" validate:"is-full-name*/
 	Phone     string    `json:"phone" binding:"required"`    /*,min=10,max=10*/
 	Password  string    `json:"password" binding:"required"` /*,min=8*/
 	Addresses []Address `json:"addresses"`
-	Balance   float32   `json:"balance"`
+	Balance   int       `json:"balance"`
 }
 
 type UserInfoRequest struct {
-	ID int8 `json:"id"`
+	ID int `json:"id"`
 }
 
 type UserLoginRequest struct {
@@ -19,8 +19,8 @@ type UserLoginRequest struct {
 }
 
 type UserEditRequest struct {
-	ID       int8    `json:"id" binding:"required"`
-	Name     string  `json:"name"`
-	Password string  `json:"password"`
-	Balance  float32 `json:"balance"`
+	ID       int    `json:"id" binding:"required"`
+	Name     string `json:"name"`
+	Password string `json:"password"`
+	Balance  int    `json:"balance"`
 }
